% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_global.R
\docType{data}
\name{BASE_I}
\alias{BASE_I}
\alias{BASE_J}
\alias{BASE_R}
\alias{BASE_L}
\alias{BASE_U}
\alias{BASE_B}
\alias{BASE_D}
\alias{BASE_P}
\alias{BASE_Q}
\alias{BASE_LIST}
\alias{BASE_C}
\title{Base patterns}
\format{
An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{list} of length 10.

An object of class \code{sfc_base} of length 1.
}
\usage{
BASE_I

BASE_J

BASE_R

BASE_L

BASE_U

BASE_B

BASE_D

BASE_P

BASE_Q

BASE_LIST

BASE_C
}
\description{
Base patterns
}
\keyword{datasets}
