% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_multiple_curves}
\alias{draw_multiple_curves}
\title{Draw multiple curves}
\usage{
draw_multiple_curves(
  ...,
  nrow = NULL,
  ncol = NULL,
  extend = TRUE,
  title = FALSE,
  closed = FALSE,
  padding = unit(0, "pt"),
  lwd = 4,
  col = NULL
)
}
\arguments{
\item{...}{A list of \code{sfc_sequence} objects or objects in its child classes. The value
can also be a list of two-column coordinate matrices.}

\item{nrow}{Number of rows in the layout.}

\item{ncol}{Number of columns in the layout.}

\item{extend}{Whether to draw the entering and leaving segments? It is only used when input is a list of \code{sfc_sequence} objects.}

\item{title}{Whether to add titles on each panel? The title is constructed in the form of \code{initial_seed|transverse_code}, e.g. \code{I|111}.
Or the value can be a vector of strings.}

\item{closed}{Whether the curves are closed? The value should be a logical vector. If it is \code{TRUE}, the last point
is connected to the first point in the curve to form a closed curve.}

\item{padding}{Space around each curve. The value should be a \code{\link[grid:grid.grob]{grid::grob}} object.}

\item{lwd}{Line width.}

\item{col}{Color for segments. If the value is \code{NULL}, it uses the "Spectral" color palettes.}
}
\description{
Draw multiple curves
}
\details{
draw_multiple_curves
This function is only for the demonstration purpose.
}
\examples{
# for all forms of curves initialized by base pattern 'R', with rotation 0, and on level 3
draw_multiple_curves(
    sfc_hilbert("R", code = c(1, 1, 1)),
    sfc_hilbert("R", code = c(1, 1, 2)),
    sfc_hilbert("R", code = c(1, 2, 1)),
    sfc_hilbert("R", code = c(1, 2, 2)),
    sfc_hilbert("R", code = c(2, 1, 1)),
    sfc_hilbert("R", code = c(2, 1, 2)),
    sfc_hilbert("R", code = c(2, 2, 1)),
    sfc_hilbert("R", code = c(2, 2, 2)),
    nrow = 2, title = TRUE)

# simply a list of sequences
draw_multiple_curves(
    sfc_sequence("IIII"),
    sfc_sequence("RRRR"),
    sfc_sequence("RRLL"),
    nrow = 1
)
}
