% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_curve.R
\name{h_curve}
\alias{h_curve}
\alias{expand_h}
\title{H-curve}
\usage{
h_curve(h, connect = c("h", "v"), iteration = 1, random = FALSE)

expand_h(h1, h2 = h1, h3 = h1, h4 = h1, connect = "vvvv")
}
\arguments{
\item{h}{The seed. It should be one of \code{\link{H0}}, \code{\link{H1}} or \code{\link{H2}}.}

\item{connect}{How the four sub-units are connected.}

\item{iteration}{Number of interations.}

\item{random}{Whether generate units randomly.}

\item{h1}{The first subunit.}

\item{h2}{The second subunit.}

\item{h3}{The third subunit.}

\item{h4}{The fourth subunit.}
}
\description{
H-curve
}
\examples{
h_curve(H0, iteration = 2) |> plot(type = "l", asp = 1)
h_curve(H1, iteration = 2) |> plot(type = "l", asp = 1)
par(mfrow = c(1, 2))
h_curve(H1, iteration = 3, random = TRUE) |> plot(type = "l", asp = 1)
h_curve(H1, iteration = 3, random = TRUE) |> plot(type = "l", asp = 1)
expand_h(H0, connect = "hvvh") |> plot(type = "l", asp = 1)
expand_h(H1, connect = "vvhh") |> plot(type = "l", asp = 1)

h1 = expand_h(H0, connect = "hhhh")
h2 = expand_h(H0, connect = "vvvv")
h3 = expand_h(H0, connect = "hvhv")
h4 = expand_h(H0, connect = "hvvh")
expand_h(h1, h2, h3, h4, connect = "vhvh") |> plot(type = "l", asp = 1)

fun = function(h, k) {
    for(i in 1:k) h = expand_h(h, connect = "vhvh")
    h
}
fun(H0, 4) |> plot(type = "l", asp = 1)
}
