% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_curve.R
\docType{data}
\name{H0}
\alias{H0}
\alias{H1}
\alias{H2}
\title{Seed sequences of the H-curve}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 16 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 16 rows and 2 columns.
}
\usage{
H0

H1

H2
}
\description{
Seed sequences of the H-curve
}
\details{
The three objects simply contain the coordinates of points on the three curves.
}
\examples{
H0
plot(H1, type = "l", asp = 1)
}
\keyword{datasets}
