% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_curve.R
\name{level1_unit_orientation,sfc_peano-method}
\alias{level1_unit_orientation,sfc_peano-method}
\alias{level1_unit_orientation}
\alias{change_level1_unit_orientation,sfc_peano-method}
\alias{change_level1_unit_orientation}
\title{Level-1 unit in the Peano curve}
\usage{
\S4method{level1_unit_orientation}{sfc_peano}(p)

\S4method{change_level1_unit_orientation}{sfc_peano}(p, to = c("horizontal", "vertical"))
}
\arguments{
\item{p}{An \code{sfc_peano} object.}

\item{to}{A string of "vertical" or "horizontal".}
}
\value{
\code{level1_unit_orientation()} returns \code{"vertical"} or \code{"horizontal"}.

\code{change_level1_unit_orientation()} returns an \code{sfc_peano} object.
}
\description{
Level-1 unit in the Peano curve
}
\examples{
p = sfc_peano("I", 111)
p2 = change_level1_unit_orientation(p)
p3 = change_level1_unit_orientation(p, "vertical")
draw_multiple_curves(p2, p3, title = FALSE)
}
