% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve.R
\name{plot_segments}
\alias{plot_segments}
\title{Plot segments}
\usage{
plot_segments(x, grid = FALSE, title = FALSE, closed = FALSE, ...)
}
\arguments{
\item{x}{A two-column matrix of coordinates of points.}

\item{grid}{Whether to add grid lines on the plot?}

\item{title}{The value should be \code{FALSE} or a string.}

\item{closed}{Whether the curve is closed?}

\item{...}{Other arguments passed to \code{\link[=sfc_grob]{sfc_grob()}}.}
}
\description{
Plot segments
}
\details{
This function is only for a quick demonstration of curves represented as two-column coordinate matrices.
}
