% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_global.R
\docType{data}
\name{SFC_RULES_HILBERT}
\alias{SFC_RULES_HILBERT}
\alias{SFC_RULES_PEANO}
\alias{SFC_RULES_MEANDER}
\title{Rules}
\format{
An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.
}
\usage{
SFC_RULES_HILBERT

SFC_RULES_PEANO

SFC_RULES_MEANDER
}
\description{
A list of pre-defined expansion rules for different curves.
}
\details{
\code{SFC_RULES_PEANO} and \code{SFC_RULES_MEANDER} contain both "flipped" expansion rules.
}
\examples{
SFC_RULES_HILBERT
SFC_RULES_PEANO
SFC_RULES_MEANDER
}
\keyword{datasets}
