% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_4x4_meander.R
\name{sfc_4x4_meander}
\alias{sfc_4x4_meander}
\alias{sfc_expand,sfc_4x4_meander-method}
\alias{draw_rules_4x4_meander}
\title{4x4 space-filling curves in meander type}
\usage{
sfc_4x4_meander(seed, code = integer(0), rot = 0L, flip = FALSE, type = 1L)

\S4method{sfc_expand}{sfc_4x4_meander}(p, code)

draw_rules_4x4_meander(type = 1)
}
\arguments{
\item{seed}{The seed sequence. In most cases, the seed sequence is a single base pattern, which can be specified as a single letter, then \code{rot} controls
the initial rotation of the base pattern. It also supports a sequence with more than one base patterns as the seed sequence. In this case,
it can be specified as a string of more than one base letters, then \code{rot} can be set to a single rotation scalar which controls the rotation of the
first letter, or a vector with the same length as the number of base letters.}

\item{code}{Transverse code, a single integer.}

\item{rot}{Rotation of the seed sequence, measured in the polar coordinate system, in degrees.}

\item{flip}{The same setting as in \code{\link[=sfc_peano]{sfc_peano()}} or \code{\link[=sfc_meander]{sfc_meander()}}.}

\item{type}{Which type of rules to use? 1 for \code{\link{SFC_RULES_MEANDER_4x4_1}} and 2 for \code{\link{SFC_RULES_MEANDER_4x4_2}}.}

\item{p}{An \code{sfc_4x4_meander} object.}
}
\description{
4x4 space-filling curves in meander type
}
\details{
It is an extension of the 3x3 Meander curves. For simplicity, it only supports \code{I/R/L} base patterns.
}
\examples{
draw_multiple_curves(
    sfc_4x4_meander("I", "11", type = 1),
    sfc_4x4_meander("I", "12", type = 1),
    sfc_4x4_meander("I", "11", type = 2),
    sfc_4x4_meander("I", "12", type = 2),
    nrow = 2
)
seed = paste(rep(paste0("R", sapply(0:10, function(i) strrep("I", i))), each = 2), collapse="")
sfc_4x4_meander(seed, 1) |> plot()
draw_rules_4x4_meander(type = 1)
draw_rules_4x4_meander(type = 2)
}
