% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_apply.R
\name{sfc_apply,sfc_nxn-method}
\alias{sfc_apply,sfc_nxn-method}
\alias{sfc_apply}
\title{Apply to every unit in the sfc_nxn curve}
\usage{
\S4method{sfc_apply}{sfc_nxn}(p, depth = 1, fun = function(x) x)
}
\arguments{
\item{p}{An \code{sfc_nxn} object.}

\item{depth}{An integer between 0 and \code{level-1} of the curve.}

\item{fun}{A function of which the argument \code{x} is a sub-unit in the curve.}
}
\description{
Apply to every unit in the sfc_nxn curve
}
\details{
The size of the square unit is determined by \code{depth}. A depth of 0 corresponds to the complete curve.
A depth of 1 corresponds to the nine first-level units, et al.
}
\examples{
p = sfc_peano("I", 111)
draw_multiple_curves(p, sfc_apply(p, 0, function(x) sfc_flip_unit(x)))
draw_multiple_curves(p, sfc_apply(p, 1, function(x) sfc_flip_unit(x)))
draw_multiple_curves(p, sfc_apply(p, 2, function(x) sfc_flip_unit(x)))

# only works on the lowest unit
peano_unit_orientation = function(u) {
    loc = sfc_segments(u)
    if(loc[1, 1] == loc[2, 1] && loc[2, 1] == loc[3, 1]) {
        "vertical"
    } else {
        "horizontal"
    }
}
p2 = sfc_apply(p, 2, function(x) {
    if(peano_unit_orientation(x) == "vertical") {
        sfc_flip_unit(x)
    } else {
        x
    }
})
draw_multiple_curves(p, p2, title = FALSE)

p3 = sfc_apply(p, 2, function(x) {
    if(peano_unit_orientation(x) == "horizontal") {
        sfc_flip_unit(x)
    } else {
        x
    }
})
draw_multiple_curves(p, p3, title = FALSE)
}
