% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_expand.R
\name{sfc_expand,sfc_hilbert-method}
\alias{sfc_expand,sfc_hilbert-method}
\alias{sfc_expand}
\alias{sfc_expand,sfc_peano-method}
\alias{sfc_expand,sfc_meander-method}
\title{Expand the curve to the next level}
\usage{
\S4method{sfc_expand}{sfc_hilbert}(p, code)

\S4method{sfc_expand}{sfc_peano}(p, code = 1)

\S4method{sfc_expand}{sfc_meander}(p, code)
}
\arguments{
\item{p}{An \code{sfc_hilbert} object.}

\item{code}{Transverse code, a single integer.}
}
\value{
An object in the same class as the input. \code{sfc_expand()} on the \code{sfc_rules} objects returns an \code{sfc_sequence} object.
}
\description{
Expand the curve to the next level
}
\details{
For the Hilbert curve and Meander curve, as long as the transverse code of the first base pattern in the sequence is determinted,
the transverse codes for other base patterns in the sequence are all determined. For the Peano curve, \code{code} is always 1.

These functions are mainly used internally.
}
\examples{
p = sfc_hilbert("I", 11)
sfc_expand(p, 2) # I|211
p = sfc_peano("I", 11)
sfc_expand(p, 2) # I|211
p = sfc_meander("I", 11)
sfc_expand(p, 2) # I|211
}
