% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_rules.R
\name{sfc_expand_by_rules,sfc_rules,factor-method}
\alias{sfc_expand_by_rules,sfc_rules,factor-method}
\alias{sfc_expand_by_rules}
\alias{sfc_expand_by_rules,sfc_rules,sfc_nxn-method}
\alias{sfc_expand_by_rules,sfc_rules,character-method}
\title{Expand a sequence}
\usage{
\S4method{sfc_expand_by_rules}{sfc_rules,factor}(
  p,
  letters,
  code = 1L,
  flip = FALSE,
  by = "Cpp",
  rot = NULL
)

\S4method{sfc_expand_by_rules}{sfc_rules,sfc_nxn}(
  p,
  letters,
  code = 1L,
  flip = FALSE,
  by = "Cpp",
  rot = NULL
)

\S4method{sfc_expand_by_rules}{sfc_rules,character}(
  p,
  letters,
  code = 1L,
  flip = FALSE,
  by = "Cpp",
  rot = NULL
)
}
\arguments{
\item{p}{An \code{sfc_rules} object.}

\item{letters}{A list of base patterns in letters, must be a factor.}

\item{code}{The transverse code.}

\item{flip}{For the Peano curve and the Meander curves, each unit can be flipped without affecting other parts in the curve. This argument
controls whether to flip the unit. Since currently it only works on the Peano curve and the Meander curve, \code{flip} should be a logical
vector of length one or length of 9. Whether it flips horizontally, vertically or against the diagonal line is automatically choosen.
The value of \code{flip} can also be a function which takes the length of \code{letters} as the only argument.}

\item{by}{Which implementation? Only for the testing purpose.}

\item{rot}{Rotations of \code{letters}.}
}
\description{
Expand a sequence
}
\examples{
sfc_expand_by_rules(SFC_RULES_HILBERT, 
    factor(c("I", "R", "L"), levels = sfc_universe(SFC_RULES_HILBERT)))
}
