% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_base.R, R/sfc_grob.R
\name{sfc_grob,sfc_base-method}
\alias{sfc_grob,sfc_base-method}
\alias{sfc_grob,sfc_sequence-method}
\alias{sfc_grob}
\alias{sfc_grob,sfc_nxn-method}
\title{The graphics object from the sfc_base object}
\usage{
\S4method{sfc_grob}{sfc_base}(p)

\S4method{sfc_grob}{sfc_sequence}(p, bases, extend = FALSE, title = FALSE, ...)

\S4method{sfc_grob}{sfc_nxn}(p, bases = p@rules@bases, extend = FALSE, title = FALSE, ...)
}
\arguments{
\item{p}{A \code{sfc_nxn} object.}

\item{bases}{A list of base patterns, consider to use \code{BASE_LIST}.}

\item{extend}{Whether to add the entering and leaving segments?}

\item{title}{Whether to add title on top of the curve?}

\item{...}{Other arguments passed to \code{\link[grid:viewport]{grid::viewport()}}.}
}
\value{
A \code{\link[grid:grid.grob]{grid::grob()}} object.

A \code{\link[grid:grid.grob]{grid::grob()}} object.
}
\description{
The graphics object from the sfc_base object

The graphics object from the sfc_nxn object
}
\examples{
sfc_grob(BASE_I)
}
