% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_sequence.R
\name{sfc_is_compatible,sfc_sequence-method}
\alias{sfc_is_compatible,sfc_sequence-method}
\alias{sfc_is_compatible}
\title{Whether two sfc_sequences objects are compatible}
\usage{
\S4method{sfc_is_compatible}{sfc_sequence}(p, p2)
}
\arguments{
\item{p}{A \code{sfc_sequence} object.}

\item{p2}{A \code{sfc_sequence} object.}
}
\description{
Whether two sfc_sequences objects are compatible
}
\details{
The function compares whether the two universe base pattern sets are identical.
}
