% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_curve.R
\name{sfc_level,sfc_nxn-method}
\alias{sfc_level,sfc_nxn-method}
\alias{sfc_level}
\title{The level of the curve}
\usage{
\S4method{sfc_level}{sfc_nxn}(p)
}
\arguments{
\item{p}{An \code{sfc_nxn} object.}
}
\value{
An integer.
}
\description{
The level of the curve
}
\examples{
p = sfc_hilbert("I", "11")
sfc_level(p)
}
