% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_curve.R, R/sfc_rules.R
\name{sfc_mode,sfc_nxn-method}
\alias{sfc_mode,sfc_nxn-method}
\alias{sfc_mode,sfc_rules-method}
\alias{sfc_mode}
\title{The mode of the curve}
\usage{
\S4method{sfc_mode}{sfc_nxn}(p)

\S4method{sfc_mode}{sfc_rules}(p)
}
\arguments{
\item{p}{The corresponding object.}
}
\value{
An integer.
}
\description{
The mode of the curve
}
\examples{
p = sfc_2x2("I", "1")
sfc_mode(p)
p = sfc_3x3_peano("I", "1")
sfc_mode(p)
sfc_mode(SFC_RULES_2x2)
sfc_mode(SFC_RULES_3x3_PEANO)
}
