% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_segments.R
\name{sfc_segments,sfc_nxn-method}
\alias{sfc_segments,sfc_nxn-method}
\alias{sfc_segments}
\alias{sfc_segments,sfc_sequence-method}
\title{Coordinates of the points on the curve}
\usage{
\S4method{sfc_segments}{sfc_nxn}(p, bases = p@rules@bases, start = c(0, 0))

\S4method{sfc_segments}{sfc_sequence}(p, bases, start = c(0, 0))
}
\arguments{
\item{p}{An \code{sfc_nxn} or \code{sfc_sequence} object.}

\item{bases}{A list of base patterns, consider to use \code{BASE_LIST}.}

\item{start}{Coordinate of the start point.}
}
\value{
A two-column matrix.
}
\description{
Coordinates of the points on the curve
}
\examples{
p = sfc_hilbert("I", "11")
loc = sfc_segments(p)
plot(loc, type = "l")
}
