% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_transform.R
\name{sfc_rotate,sfc_sequence-method}
\alias{sfc_rotate,sfc_sequence-method}
\alias{sfc_rotate}
\alias{sfc_hflip,sfc_sequence-method}
\alias{sfc_hflip}
\alias{sfc_vflip,sfc_sequence-method}
\alias{sfc_vflip}
\alias{sfc_dflip,sfc_sequence-method}
\alias{sfc_dflip}
\alias{sfc_reverse,sfc_sequence-method}
\alias{sfc_reverse}
\title{Transformations of a sequence}
\usage{
\S4method{sfc_rotate}{sfc_sequence}(p, rot)

\S4method{sfc_hflip}{sfc_sequence}(p, fix_ends = FALSE)

\S4method{sfc_vflip}{sfc_sequence}(p, fix_ends = FALSE)

\S4method{sfc_dflip}{sfc_sequence}(p, slop = 1, fix_ends = FALSE)

\S4method{sfc_reverse}{sfc_sequence}(p)
}
\arguments{
\item{p}{An \code{sfc_sequence} object.}

\item{rot}{Rotation measured in the polar coordinate system, in degree.}

\item{fix_ends}{Whether to keep the rotations of the two end base patterns?}

\item{slop}{Slop of the diagonal.}
}
\description{
Transformations of a sequence
}
\details{
\itemize{
\item \code{sfc_rotate()} rotates each base pattern.
\item \code{sfc_hflip()} flips a sequence horizontally.
\item \code{sfc_vflip()} flips a sequence vertically.
\item \code{sfc_dflio()} flips a sequence against a diagonal line.
}
}
