% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_rules.R, R/sfc_sequence.R
\name{sfc_universe,sfc_rules-method}
\alias{sfc_universe,sfc_rules-method}
\alias{sfc_universe}
\alias{sfc_universe,sfc_sequence-method}
\title{The universe pattern set}
\usage{
\S4method{sfc_universe}{sfc_rules}(p)

\S4method{sfc_universe}{sfc_sequence}(p)
}
\arguments{
\item{p}{The corresponding object.}
}
\value{
A vector of letters
}
\description{
The universe pattern set
}
\examples{
sfc_universe(SFC_RULES_HILBERT)
sfc_universe(SFC_RULES_PEANO)
}
