% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_validate.R
\name{sfc_validate,sfc_sequence-method}
\alias{sfc_validate,sfc_sequence-method}
\alias{sfc_validate}
\alias{sfc_validate,character-method}
\title{Validate the sequence}
\usage{
\S4method{sfc_validate}{sfc_sequence}(p, by = "sfc_hilbert")

\S4method{sfc_validate}{character}(p, by = "sfc_hilbert")
}
\arguments{
\item{p}{An \code{sfc_sequence} object or a character string. If it is a character string, rotations are all taken as zeros.}

\item{by}{One of \code{sfc_hilbert}, \code{sfc_peano} and \code{sfc_meander}.}
}
\description{
Validate the sequence
}
\details{
It is mainly used to validate a seed sequence whether they follow the forward-left-right rule.
}
\examples{
try(sfc_validate("LLLLL"))
try(sfc_validate(sfc_sequence("IIIII", rot = c(0, 90, 180, 270, 0), 
        universe = sfc_universe(SFC_RULES_HILBERT))))
}
