% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class.R
\docType{package}
\name{sfcurve-package}
\alias{sfcurve}
\alias{sfcurve-package}
\title{sfcurve: 2x2, 3x3 and nxn Space-Filling Curves}
\description{
Implementation of all possible forms of 2x2 and 3x3 space-filling curves, i.e., the generalized forms of the Hilbert curve \url{https://en.wikipedia.org/wiki/Hilbert_curve}, the Peano curve \url{https://en.wikipedia.org/wiki/Peano_curve} and the Peano curve in the meander type (Figure 5 in \url{https://eudml.org/doc/141086}). It generates nxn curves expanded from specific level-1 units. It also implements the H-curve and the three-dimensional Hilbert curve.
}
\details{
Please go to the package's website: \url{https://jokergoo.github.io/sfcurve/} for more details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jokergoo/sfcurve}
  \item \url{https://jokergoo.github.io/sfcurve/}
}

}
\author{
\strong{Maintainer}: Zuguang Gu \email{z.gu@dkfz.de} (\href{https://orcid.org/0000-0002-7395-8709}{ORCID})

}
\keyword{internal}
