% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_base.R, R/sfc_curve.R, R/sfc_rules.R,
%   R/sfc_sequence.R
\name{show,sfc_base-method}
\alias{show,sfc_base-method}
\alias{show,sfc_nxn-method}
\alias{show,sfc_rules-method}
\alias{show,sfc_sequence-method}
\title{Print the object}
\usage{
\S4method{show}{sfc_base}(object)

\S4method{show}{sfc_nxn}(object)

\S4method{show}{sfc_rules}(object)

\S4method{show}{sfc_sequence}(object)
}
\arguments{
\item{object}{The corresponding object.}
}
\description{
Print the object
}
