% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transverse_path.R
\name{all_transverse_paths}
\alias{all_transverse_paths}
\alias{get_one_transverse_path}
\alias{plot_transverse_paths}
\title{All transverse paths of a sequence}
\usage{
all_transverse_paths(rules, p)

get_one_transverse_path(rules, p)

plot_transverse_paths(rules, p, paths, type = 0)
}
\arguments{
\item{rules}{An \code{sfc_rules} object.}

\item{p}{An \code{sfc_sequence} sequence. \code{p} and \code{rules} should have the same universe base set.}

\item{paths}{The object returned from \code{all_transverse_paths()}.}

\item{type}{If the value is 1, it highlights the paths only via 1-1/2-2 corners. If the value is 2, it highlights the paths
only via 1-2/2-1 corners.}
}
\description{
All transverse paths of a sequence
}
\details{
Given an input sequence with rotations, list all combinations of transverse code for each letter in p.

\code{get_one_transverse_path()} returns one random transverse path.
}
\examples{
p = sfc_3x3_combined("I", level = 1)
get_one_transverse_path(SFC_RULES_3x3_COMBINED, p)
get_one_transverse_path(SFC_RULES_3x3_COMBINED, p)
get_one_transverse_path(SFC_RULES_3x3_COMBINED, p)
get_one_transverse_path(SFC_RULES_3x3_COMBINED, p)
p = SFC_RULES_3x3_COMBINED@rules$I[[1]]
pl = all_transverse_paths(SFC_RULES_3x3_COMBINED, p)
plot_transverse_paths(SFC_RULES_3x3_COMBINED, p, pl)
plot_transverse_paths(SFC_RULES_3x3_COMBINED, p, pl, type = 1)
plot_transverse_paths(SFC_RULES_3x3_COMBINED, p, pl, type = 2)
}
