% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_global.R
\docType{data}
\name{BASE_I}
\alias{BASE_I}
\alias{BASE_J}
\alias{BASE_R}
\alias{BASE_L}
\alias{BASE_U}
\alias{BASE_B}
\alias{BASE_D}
\alias{BASE_P}
\alias{BASE_Q}
\alias{BASE_C}
\alias{BASE_LIST}
\title{Base patterns}
\format{
An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{sfc_base} of length 1.

An object of class \code{list} of length 10.
}
\usage{
BASE_I

BASE_J

BASE_R

BASE_L

BASE_U

BASE_B

BASE_D

BASE_P

BASE_Q

BASE_C

BASE_LIST
}
\description{
A list of pre-defined base patterns. See the \strong{Examples} section.
}
\details{
\code{BASE_I} and \code{BASE_J} are identical. They are only used to distinguish the two
"going forward" patterns in the 3x3 mode for the Peano curves, i.e. bottom-left to top-right, and bottom-right
to top-left.
}
\examples{
BASE_I
BASE_J
BASE_R
BASE_L
BASE_U
BASE_B
BASE_D
BASE_P
BASE_Q
BASE_C
draw_multiple_curves(
    BASE_I, BASE_J, BASE_R, BASE_L, BASE_U,
    BASE_B, BASE_D, BASE_P, BASE_Q, BASE_C,
    nrow = 1
)
}
\keyword{datasets}
