% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_multiple_curves}
\alias{draw_multiple_curves}
\title{Draw multiple curves}
\usage{
draw_multiple_curves(..., nrow = 1, ncol = NULL, extend = TRUE, title = TRUE)
}
\arguments{
\item{...}{A list of \verb{sfc_*} objects.}

\item{nrow}{Number of rows in the layout.}

\item{ncol}{Number of columns in the layout.}

\item{extend}{Whether to draw the entering and leaving segments?}

\item{title}{Whether to add titles on each panel? The title is constructed in the form of \code{initial_seed|transverse_code}, e.g. \code{I|111}.}
}
\description{
Draw multiple curves
}
\examples{
# for all forms of curves initialized by 'R', rotation 0, and in level 3
draw_multiple_curves(
    sfc_hilbert("R", code = c(1, 1, 1)),
    sfc_hilbert("R", code = c(1, 1, 2)),
    sfc_hilbert("R", code = c(1, 2, 1)),
    sfc_hilbert("R", code = c(1, 2, 2)),
    sfc_hilbert("R", code = c(2, 1, 1)),
    sfc_hilbert("R", code = c(2, 1, 2)),
    sfc_hilbert("R", code = c(2, 2, 1)),
    sfc_hilbert("R", code = c(2, 2, 2)),
    nrow = 2, title = TRUE)

# simply a list of sequences
draw_multiple_curves(
    sfc_sequence("IIII"),
    sfc_sequence("RRRR"),
    sfc_sequence("RRLL"),
    nrow = 1
)
}
