% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_rules.R
\name{draw_rules_hilbert}
\alias{draw_rules_hilbert}
\alias{draw_rules_peano}
\alias{draw_rules_meander}
\title{Draw the expansion rules}
\usage{
draw_rules_hilbert()

draw_rules_peano(flip = FALSE)

draw_rules_meander(flip = FALSE)
}
\arguments{
\item{flip}{Whether to use the "flipped" rules? For the Peano curve and the Meander curve, there is also a "fliiped" version
of the curve expansion rules. See the vignettes for details.}
}
\description{
Draw the expansion rules
}
\details{
The expansion rules define how the curve is expanded from level-0 to level-1.
}
\examples{
draw_rules_hilbert()
# the units in the main rules of the Peano curve are vertical
draw_rules_peano()
# the units in the flipped rules of the Peano curve are horizontal
draw_rules_peano(flip = TRUE)
# the units in the main rules of the Meander curve are "forward"
# i.e. the direction of the "wave" is the same as the direction of the curve
draw_rules_meander()
# the units in the flipped rules of the Meander curve are "backward"
draw_rules_meander(flip = TRUE)
}
