% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_rules.R
\name{draw_rules_hilbert}
\alias{draw_rules_hilbert}
\alias{draw_rules_peano}
\alias{draw_rules_meander}
\title{Draw the expansion rules}
\usage{
draw_rules_hilbert()

draw_rules_peano(flip = FALSE)

draw_rules_meander(flip = FALSE)
}
\arguments{
\item{flip}{Whether to use the "flipped" rules? For the Peano curve and the Meander curve, there is also a "fliiped" version
of curve expansion rules. See the vignettes for details.}
}
\description{
Draw the expansion rules
}
\examples{
draw_rules_hilbert()
draw_rules_peano()
draw_rules_peano(flip = TRUE)
draw_rules_meander()
draw_rules_meander(flip = TRUE)
}
