% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_curve.R
\name{level1_unit_orientation,sfc_3x3_peano-method}
\alias{level1_unit_orientation,sfc_3x3_peano-method}
\alias{level1_unit_orientation}
\alias{change_level1_unit_orientation,sfc_3x3_peano-method}
\alias{change_level1_unit_orientation}
\title{Level-1 unit in the Peano curve}
\usage{
\S4method{level1_unit_orientation}{sfc_3x3_peano}(p)

\S4method{change_level1_unit_orientation}{sfc_3x3_peano}(p, to = c("horizontal", "vertical"))
}
\arguments{
\item{p}{For \code{level1_unit_orientation()}, it is an \code{sfc_3x3_peano} unit on level-1.
For \code{change_level1_unit_orientation()}, it is a normal \code{sfc_3x3_peano} object.}

\item{to}{A string of "vertical" or "horizontal".}
}
\value{
\code{level1_unit_orientation()} returns \code{"vertical"} or \code{"horizontal"}.

\code{change_level1_unit_orientation()} returns an \code{sfc_3x3_peano} object.
}
\description{
Level-1 unit in the Peano curve
}
\details{
\code{level1_unit_orientation()} is normally used inside \code{\link[=sfc_apply]{sfc_apply()}}. \code{change_level1_unit_orientation()}
changes all level-1 units of a Peano curves simultaneously.
}
\examples{
p = sfc_3x3_peano("I", 111)
level1_unit_orientation(p[1:9, TRUE])
level1_unit_orientation(p[1:9 + 27, TRUE])
p2 = change_level1_unit_orientation(p)
p3 = change_level1_unit_orientation(p, "vertical")
draw_multiple_curves(p, p2, p3, 
    title = c("original", "all horizontal", "all vertical"), nrow = 1)
}
