% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_global.R
\docType{data}
\name{SFC_RULES_2x2}
\alias{SFC_RULES_2x2}
\alias{SFC_RULES_3x3_PEANO}
\alias{SFC_RULES_3x3_MEANDER}
\alias{SFC_RULES_3x3_COMBINED}
\alias{SFC_RULES_4x4_MEANDER_1}
\alias{SFC_RULES_4x4_MEANDER_2}
\title{Rules}
\format{
An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.
}
\usage{
SFC_RULES_2x2

SFC_RULES_3x3_PEANO

SFC_RULES_3x3_MEANDER

SFC_RULES_3x3_COMBINED

SFC_RULES_4x4_MEANDER_1

SFC_RULES_4x4_MEANDER_2
}
\description{
A list of pre-defined expansion rules for different curves.
}
\details{
\code{SFC_RULES_3x3_PEANO}, \code{SFC_RULES_3x3_MEANDER} and \code{SFC_RULES_3x3_COMBINED}, \code{SFC_RULES_MEANDER_4x4_1}, \code{SFC_RULES_MEANDER_4x4_2} also contain the "flipped" expansion rules.

\code{SFC_RULES_3x3_COMBINED} is a combination of \code{SFC_RULES_3x3_PEANO} and \code{SFC_RULES_3x3_MEANDER} where
in \code{SFC_RULES_3x3_PEANO}, \code{J} is replaced by its original pattern \code{I}.

\code{SFC_RULES_4x4_MEANDER_1} and \code{SFC_RULES_4x4_MEANDER_2} are extension rules of Meander curves (3x3)
on the 4x4 mode It is only for the demonstration purpose, thus only \code{I/R/L} are supported.
}
\examples{
SFC_RULES_2x2
SFC_RULES_3x3_PEANO
SFC_RULES_3x3_MEANDER
SFC_RULES_3x3_COMBINED
SFC_RULES_4x4_MEANDER_1
SFC_RULES_4x4_MEANDER_2
}
\keyword{datasets}
