% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_global.R
\docType{data}
\name{SFC_RULES_HILBERT}
\alias{SFC_RULES_HILBERT}
\alias{SFC_RULES_PEANO}
\alias{SFC_RULES_PEANO_FLIP}
\alias{SFC_RULES_MEANDER}
\alias{SFC_RULES_MEANDER_FLIP}
\title{Rules}
\format{
An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.

An object of class \code{sfc_rules} of length 1.
}
\usage{
SFC_RULES_HILBERT

SFC_RULES_PEANO

SFC_RULES_PEANO_FLIP

SFC_RULES_MEANDER

SFC_RULES_MEANDER_FLIP
}
\description{
Rules
}
\keyword{datasets}
