% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_expand.R, R/sfc_rules.R
\name{sfc_expand,sfc_hilbert-method}
\alias{sfc_expand,sfc_hilbert-method}
\alias{sfc_expand}
\alias{sfc_expand,sfc_peano-method}
\alias{sfc_expand,sfc_meander-method}
\alias{sfc_expand,sfc_rules-method}
\title{Expand the curve to the next level}
\usage{
\S4method{sfc_expand}{sfc_hilbert}(p, code)

\S4method{sfc_expand}{sfc_peano}(p, code = 1)

\S4method{sfc_expand}{sfc_meander}(p, code)

\S4method{sfc_expand}{sfc_rules}(p, letters, code = 1L, by = "Cpp")
}
\arguments{
\item{p}{An \code{sfc_rules} object.}

\item{code}{The transverse code.}

\item{letters}{A list patterns, must be a factor.}

\item{by}{Which implementation?}
}
\value{
An \code{sfc_hilbert} object.
}
\description{
Expand the curve to the next level

Expand the patterns to the next level
}
\details{
For the Hilbert curve and Meander curve, as long as the transverse code of the first base pattern in the sequence is determinted,
the transverse codes for other base patterns in the sequence are all determined. For the Peano curve, \code{code} is always 1.

It is mainly used internally.
}
\examples{
p = sfc_hilbert("I", 11)
sfc_expand(p, 1)
p = sfc_peano("I", 11)
sfc_expand(p, 1)
p = sfc_meander("I", 11)
sfc_expand(p, 1)
sfc_expand(SFC_RULES_HILBERT, 
    factor(c("I", "R", "L"), levels = sfc_universe(SFC_RULES_HILBERT)))
}
