% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_flip_unit.R
\name{unit_orientation}
\alias{unit_orientation}
\alias{sfc_flip_unit,sfc_nxn-method}
\alias{sfc_flip_unit}
\alias{sfc_flip_unit,sfc_unit-method}
\title{Flip units}
\usage{
unit_orientation(p, index = "")

\S4method{sfc_flip_unit}{sfc_nxn}(p, index = "", to = NULL)

\S4method{sfc_flip_unit}{sfc_unit}(p, bases)
}
\arguments{
\item{p}{The corresponding object.}

\item{index}{A string of digits representing the path on the hierarchy of the curve. The left
side corresponds to the lower level and the right side corresponds to the high level
in the curve. For the Hilbert curve, the digits can only be 1-4, and for the Peano and
Meander curves, the digites can be 1-9. See examples in \code{\link[=sfc_index]{sfc_index()}}. The value can also
be a vector where each flipping is applied in sequence.}

\item{to}{The orientation to flip to. If the specified unit already has such orientation, the function returns
the original curve.}

\item{bases}{Normally use \code{\link{BASE_LIST}}.}
}
\description{
Flip units
}
\details{
An unit in the curve is represented as a square block (\verb{2^k x 2^k} for the Hilbert curve and \verb{3^k x 3^k} for the Peano and Meander curves, \code{k} between 1 and the level of the curve).
In the Hilbert curve, if an unit can be flipped, it is symmetric, thus flipping in the Hilbert curve does not change its form.
The flipping is mainly applied on the Peano curve and the Meander curves. Peano curve only allows flippings by the diagonals and the Meander
curve only allows flipping horizontally or vertically. The type of flipping is choosen automatically in the function.
}
\examples{
p = sfc_meander("I", 11)
draw_multiple_curves(
    p, 
    sfc_flip_unit(p, "1"), # bottom left
    sfc_flip_unit(p, "2"), # bottom middle
    sfc_flip_unit(p, "3"), # bottom right
    title = FALSE, nrow = 2)

p = sfc_peano("I", level = 3)
draw_multiple_curves(
    p, 
    sfc_flip_unit(p, ""),
    sfc_flip_unit(p, "2"),
    sfc_flip_unit(p, "21"),
    title = FALSE, nrow = 2)

p = sfc_peano("I", level = 2)
draw_multiple_curves(p, 
    sfc_flip_unit(p, c("4", "7")),
    sfc_flip_unit(p, c("1", "2", "3", "5", "6", "8", "9")),
    title = FALSE, nrow = 1)
}
