% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_index.R
\name{[.sfc_nxn}
\alias{[.sfc_nxn}
\alias{sfc_index,sfc_nxn-method}
\alias{sfc_index}
\title{Subunit in the curve}
\usage{
\method{[}{sfc_nxn}(x, i, j, ..., drop = TRUE)

\S4method{sfc_index}{sfc_nxn}(p, index = "")
}
\arguments{
\item{x}{An \code{sfc_nxn} object.}

\item{i}{Numeric index or a character index representing the hierarchy of the subunit in the curve.}

\item{j}{A value of \code{TRUE} or \code{FALSE} that controls whether to keep the \code{sfc_nxn} class or degenerate to the \code{sfc_sequence} class.}

\item{...}{Ignore.}

\item{drop}{A value of \code{TRUE} or \code{FALSE} that controls whether to keep the \code{sfc_nxn} class or degenerate to the \code{sfc_sequence} class.}

\item{p}{An \code{sfc_nxn} object.}

\item{index}{A string of digits representing the path on the hierarchy of the curve. The left side
corresponds to the top level and the right side corresponds to the bottom level on the curve. For the
Hilbert curve, the digits can only be 1-4, and for the Peano and Meander curves, the digites can be 1-9.
Digits should be separated by non-number characters, see examples.}
}
\description{
Subunit in the curve
}
\details{
\code{sfc_index()} only works on square curves (i.e. a curve with a single base letter as seed.)
}
\examples{
p = sfc_hilbert("I", "11111")
p["3:2:1"]
p["3:2:1", TRUE]
# only for testing
p = sfc_hilbert("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "3")
sfcurve:::test_sfc_index(p, "3:2")
sfcurve:::test_sfc_index(p, "3:2:1")
sfcurve:::test_sfc_index(p, "3:2:1:1")
par(om)

p = sfc_meander("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "7")
sfcurve:::test_sfc_index(p, "7:5")
sfcurve:::test_sfc_index(p, "7:5:9")
sfcurve:::test_sfc_index(p, "7:5:9:2")
par(om)
}
