% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_nxn.R
\name{[.sfc_nxn}
\alias{[.sfc_nxn}
\alias{sfc_index,sfc_nxn-method}
\alias{sfc_index}
\alias{unit_orientation}
\alias{sfc_flip_unit,sfc_nxn-method}
\alias{sfc_flip_unit}
\alias{sfc_flip_unit,sfc_unit-method}
\title{Units in the curve}
\usage{
\method{[}{sfc_nxn}(x, i, j, ..., drop = TRUE)

\S4method{sfc_index}{sfc_nxn}(p, index = "")

unit_orientation(p, index = "")

\S4method{sfc_flip_unit}{sfc_nxn}(p, index = "")

\S4method{sfc_flip_unit}{sfc_unit}(p, bases)
}
\arguments{
\item{x}{An \code{sfc_nxn} object.}

\item{i}{Numeric index or a character index representing the hierarchy of the subunit in the curve.}

\item{j}{A value of \code{TRUE} or \code{FALSE} that controls whether to keep the \code{sfc_nxn} class or degenerate to the \code{sfc_sequence} class.}

\item{...}{Ignore.}

\item{drop}{Ignore.}

\item{p}{An \code{sfc_nxn} object.}

\item{index}{A string of digits representing the path on the hierarchy of the curve. The left side
corresponds to the lower level and the right side corresponds to the high level in the curve. For the
Hilbert curve, the digits can only be 1-4, and for the Peano and Meander curves, the digites can be 1-9.}

\item{bases}{Normally use \code{\link{BASE_LIST}}.}
}
\description{
Units in the curve
}
\details{
An unit in the curve is represented as a square block (\verb{2^k x 2^k} for the Hilbert curve and \verb{3^k x 3^k} for the Peano and Meander curves).
In the Hilbert curve, if an unit can be flipped, it is symmetric, thus flipping in the Hilbert curve does not change its form.
The flipping is mainly applied in the Peano curve and the Meander curves. Peano curve only allows flippings by the diagonals and the Meander
curve only allows flipping horizontally or vertically. The type of flipping is choosen automatically in the function.
}
\examples{
p = sfc_hilbert("I", "11111")
p["321"]
p["321", TRUE]
# only for testing
p = sfc_hilbert("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "3")
sfcurve:::test_sfc_index(p, "32")
sfcurve:::test_sfc_index(p, "321")
sfcurve:::test_sfc_index(p, "3211")
par(om)

p = sfc_meander("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "7")
sfcurve:::test_sfc_index(p, "75")
sfcurve:::test_sfc_index(p, "759")
sfcurve:::test_sfc_index(p, "7592")
par(om)
p = sfc_meander("I", 11)
draw_multiple_curves(p, 
    sfc_flip_unit(p, "1"),
    sfc_flip_unit(p, "2"),
    sfc_flip_unit(p, "3"),
    title = FALSE, nrow = 1)
}
