% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_nxn.R
\name{sfc_index,sfc_nxn-method}
\alias{sfc_index,sfc_nxn-method}
\alias{sfc_index}
\alias{unit_orientation}
\alias{sfc_flip_unit,sfc_nxn-method}
\alias{sfc_flip_unit}
\alias{sfc_flip_unit,sfc_unit-method}
\title{Units in the curve}
\usage{
\S4method{sfc_index}{sfc_nxn}(p, index = "")

unit_orientation(p, index = "")

\S4method{sfc_flip_unit}{sfc_nxn}(p, index = "")

\S4method{sfc_flip_unit}{sfc_unit}(p, bases)
}
\arguments{
\item{p}{An \code{sfc_nxn} object or other corresponding object.}

\item{index}{A string representing the path of transverse the hierarchy of the curve. The left side
corresponds to the lower level and the right side corresponds to the high level in the curve. For the
Hilbert curve, the digits can only be 1-4, and for the Peano and Meander curves, the digites can be 1-9.}

\item{bases}{Noramlly use \code{\link{BASE_LIST}}.}
}
\description{
Units in the curve
}
\details{
A unit in the curve is a square block (\verb{2^k x 2^k} for the Hilbert curve and \verb{3^k x 3^k} for the Peano and Meander curves).
In the Hilbert curve, if an unit can be fliped, it is symmetric, thus flipping in the Hilbert curve will not change its form.
The flipping is mainly used in the Peano and Meander curves. Peano curve only allows flippings by the diagonals and the Meander
curve only allows flipping horizontally or vertically. The type of flipping is choosen automatically in the function.
}
\examples{
# only for testing
p = sfc_hilbert("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "3")
sfcurve:::test_sfc_index(p, "32")
sfcurve:::test_sfc_index(p, "321")
sfcurve:::test_sfc_index(p, "3211")
par(om)

p = sfc_meander("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "7")
sfcurve:::test_sfc_index(p, "75")
sfcurve:::test_sfc_index(p, "759")
sfcurve:::test_sfc_index(p, "7592")
par(om)
p = sfc_meander("I", 11)
draw_multiple_curves(p, 
    sfc_flip_unit(p, "1"),
    sfc_flip_unit(p, "2"),
    sfc_flip_unit(p, "3"),
    title = FALSE, nrow = 1)
}
