% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_index.R
\name{[.sfc_nxn}
\alias{[.sfc_nxn}
\alias{sfc_index,sfc_nxn-method}
\alias{sfc_index}
\title{Subunit in the curve}
\usage{
\method{[}{sfc_nxn}(x, i, j, ..., drop = TRUE)

\S4method{sfc_index}{sfc_nxn}(p, index = "")
}
\arguments{
\item{x}{An \code{sfc_nxn} object.}

\item{i}{Numeric index or a character index representing the hierarchy of the subunit in the curve.}

\item{j}{A value of \code{TRUE} or \code{FALSE} that controls whether to keep the \code{sfc_nxn} class or degenerate to the \code{sfc_sequence} class.}

\item{...}{Ignore.}

\item{drop}{Ignore.}

\item{p}{An \code{sfc_nxn} object.}

\item{index}{A string of digits representing the path on the hierarchy of the curve. The left side
corresponds to the lower level and the right side corresponds to the high level in the curve. For the
Hilbert curve, the digits can only be 1-4, and for the Peano and Meander curves, the digites can be 1-9.}
}
\description{
Subunit in the curve
}
\examples{
p = sfc_hilbert("I", "11111")
p["321"]
p["321", TRUE]
# only for testing
p = sfc_hilbert("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "3")
sfcurve:::test_sfc_index(p, "32")
sfcurve:::test_sfc_index(p, "321")
sfcurve:::test_sfc_index(p, "3211")
par(om)

p = sfc_meander("I", "11111")
om = par(no.readonly = TRUE)
par(mfrow = c(2, 2))
sfcurve:::test_sfc_index(p, "7")
sfcurve:::test_sfc_index(p, "75")
sfcurve:::test_sfc_index(p, "759")
sfcurve:::test_sfc_index(p, "7592")
par(om)
}
