% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_sequence.R
\name{sfc_is_compatible,sfc_sequence,sfc_sequence-method}
\alias{sfc_is_compatible,sfc_sequence,sfc_sequence-method}
\alias{sfc_is_compatible}
\alias{sfc_is_compatible,sfc_sequence,sfc_rules-method}
\alias{sfc_is_compatible,sfc_rules,sfc_sequence-method}
\title{Whether two sfc_sequence objects are compatible}
\usage{
\S4method{sfc_is_compatible}{sfc_sequence,sfc_sequence}(p1, p2)

\S4method{sfc_is_compatible}{sfc_sequence,sfc_rules}(p1, p2)

\S4method{sfc_is_compatible}{sfc_rules,sfc_sequence}(p1, p2)
}
\arguments{
\item{p1}{An \code{sfc_sequence} object.}

\item{p2}{An \code{sfc_sequence} object.}
}
\description{
Whether two sfc_sequence objects are compatible
}
\details{
The function compares whether the two universe base pattern sets are identical.
Note the order of universe base patterns should also be identical.
}
\examples{
p1 = sfc_hilbert("I")
p2 = sfc_hilbert("R")
sfc_is_compatible(p1, p2)

p1 = sfc_sequence("ABC")
p2 = sfc_sequence("DEF")
sfc_is_compatible(p1, p2)
}
