% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_shape.R
\name{sfc_shape,sfc_2x2-method}
\alias{sfc_shape,sfc_2x2-method}
\alias{sfc_shape}
\alias{all_2x2_shapes}
\title{Shape of the curve}
\usage{
\S4method{sfc_shape}{sfc_2x2}(p)

all_2x2_shapes(level = 2)
}
\arguments{
\item{p}{An \code{sfc_2x2} object.}

\item{level}{Level of the Hilbert curve.}
}
\value{
\code{sfc_shape()} returns a two-column data frame of the xy-coordinates of the shape curve.

\code{all_2x2_shapes()} returns a list of \code{n} two-column data frames where each data frame corresponds to
the xy-coordnates of the corresponding shape curve.
}
\description{
Shape of the curve
}
\details{
The shape of the curve is defined as a form of the curve without considering rotation,
flipping (reflection) nor reversing.

The process of selecting the shape segment of the curve denoted as \code{P} is:
\enumerate{
\item The entry-point should locate in the bottom left subunit and the exit-point should
locate in the bottom right subunit. We try the four rotations (0, 90, 180, 270), and
the four rotations on the horizontally flipped curve. Once we find the transformed curve
that satisfies this criterion, we term it as \code{P2}.
\item We also generate \code{P3} which is a horizontally flipped version of \code{rev(P2)}.
\item We compare the first point \code{p} of \code{P2} and \code{P3}, and select the one whose \code{p} has the smaller
x-coordinate (i.e. more to the left of the curve). If the x-coordinates of \code{p} are the same in
\code{P2} and \code{P3}, we select the one whose \code{p} has the smaller y-coordinate.
}
}
\examples{
sl = all_2x2_shapes(2)
draw_multiple_curves(list = sl, lwd = 2, col = "black")
sl = all_2x2_shapes(3)
draw_multiple_curves(list = sl, lwd = 2, col = "black")
}
