% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_validate.R
\name{sfc_validate,sfc_sequence-method}
\alias{sfc_validate,sfc_sequence-method}
\alias{sfc_validate}
\title{Validate the sequence}
\usage{
\S4method{sfc_validate}{sfc_sequence}(p, by = "sfc_hilbert")
}
\arguments{
\item{p}{An \code{sfc_sequence} object.}

\item{by}{One of \code{sfc_hilbert}, \code{sfc_peano} and \code{sfc_meander}.}
}
\description{
Validate the sequence
}
\details{
It is mainly used to validate a seed sequence whether they follow the forward-left-right growth rule
}
