% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_curve.R
\name{sfc_hilbert}
\alias{sfc_hilbert}
\alias{sfc_peano}
\alias{sfc_meander}
\title{Create space-filling curves}
\usage{
sfc_hilbert(seed, code = integer(0), rot = 0L)

sfc_peano(seed, code = integer(0), rot = 0L, flip_rules = FALSE)

sfc_meander(seed, code = integer(0), rot = 0L, flip_rules = FALSE)
}
\arguments{
\item{seed}{The seed sequence. In most cases, the seed sequence is a single base pattern, which can be specified as a single letter, then \code{rot} controls
the initial rotation of the base pattern. It also supports a sequence with more than one base patterns as the seed sequence. In this case,
it can be specified as a string of more than one base letters, then \code{rot} can be set to a single rotation scale which controls the rotation of the
first letter, or a vector with the same length as the number of base letters.}

\item{code}{A vector of the transverse code. The left side corresponds to the lower levels of the curve and the right side corresponds to the higher level of the curve.
The value can be set as a vector e.g. \code{c(1, 2, 1)}, or as a string e.g. \code{"121"}, or as a number e.g. \code{121}.}

\item{rot}{Rotation of the seed sequence, measured in the polar coordinate system, in degrees.}

\item{flip_rules}{Whether to usethe "flipped" rules? For the Peano curve and the Meander curve, there is also a "fliiped" version
of curve expansion rules. See the vignettes for details.}
}
\value{
\code{sfc_hilbert()} returns an \code{sfc_hilbert} object.
\code{sfc_peano()} returns an \code{sfc_peano} object.
\code{sfc_meander()} returns an \code{sfc_meander} object.
}
\description{
Create space-filling curves
}
\details{
\code{sfc_hilbert()} generates the Hilbert curve from the seed sequence.
\code{sfc_peano()} generates the Peano curve from the seed sequence.
\code{sfc_meander()} generates the Meander curve from the seed sequence.
}
\examples{
sfc_hilbert("I", "111") |> plot()
sfc_hilbert("I", "111", rot = 90) |> plot()
sfc_hilbert("IR", "111", rot = 90) |> plot()

sfc_peano("I", "111") |> plot()
sfc_peano("IJ", "111") |> plot()

sfc_meander("I", "111") |> plot()
sfc_meander("IR", "111") |> plot()
}
