% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve.R
\name{hilbert_curve}
\alias{hilbert_curve}
\alias{moore_curve}
\alias{beta_omega_curve}
\alias{peano_curve}
\alias{meander_curve}
\title{Various curves in their standard forms}
\usage{
hilbert_curve(level = 2)

moore_curve(level = 2)

beta_omega_curve(level = 2)

peano_curve(level = 2)

meander_curve(level = 2)
}
\arguments{
\item{level}{Level of the curve.}
}
\value{
A two-column matrix of the coordinates of points on the curve.
}
\description{
Various curves in their standard forms
}
\details{
These are just special forms of \code{\link[=sfc_hilbert]{sfc_hilbert()}}, \code{\link[=sfc_peano]{sfc_peano()}} and \code{\link[=sfc_meander]{sfc_meander()}}
}
\examples{
hilbert_curve(2)
draw_multiple_curves(
    hilbert_curve(3),
    hilbert_curve(4)
)
draw_multiple_curves(
    moore_curve(3),
    moore_curve(4)
)
draw_multiple_curves(
    beta_omega_curve(3),
    beta_omega_curve(4)
)
draw_multiple_curves(
    peano_curve(2),
    peano_curve(3)
)
draw_multiple_curves(
    meander_curve(2),
    meander_curve(3)
)
}
