% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_sequence.R
\name{[.sfc_sequence}
\alias{[.sfc_sequence}
\alias{[<-.sfc_sequence}
\alias{length.sfc_sequence}
\alias{c.sfc_sequence}
\title{Utility functions}
\usage{
\method{[}{sfc_sequence}(x, i)

\method{[}{sfc_sequence}(x, i) <- value

\method{length}{sfc_sequence}(x)

\method{c}{sfc_sequence}(...)
}
\arguments{
\item{x}{An \code{sfc_sequence} object.}

\item{i}{Index.}

\item{value}{An \code{sfc_sequence} object.}

\item{...}{A list of \code{sfc_sequence} objects or other arguments.}
}
\description{
Utility functions
}
\details{
For efficiency, \code{c.sfc_sequence()} does not check whether the input \code{sfc_sequence} objects are compatible.
}
